Путь пальцем проложи…

+++ src/main/java/faang/school/godbless/BJS2_23889/City.java
@@ -2,12 +2,11 @@ package faang.school.godbless.BJS2_23889;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @AllArgsConstructor
 @Getter
 public class City {
     private final String name;
     private final Location location;
-    private double distanceToWitcher;
+    private int distanceToWitcher;
 }
